Update main and update some files

+++ include/window.h
@@ -4,6 +4,6 @@
 #include "common.h"
 
 /* Aplica ventana in-place sobre N muestras */
-void window_apply(double *x, int N, win_t wtype);
+void window_apply(float *x, int N, win_t wtype);
 
 #endif /* WINDOW_H */
